requirements.txt
face_recognition
opencv-python
mysql-connector-python